Make PatternSet more memory efficient
Each pattern set contains four sets for
includes, excludes, includeSpecs and excludeSpecs.
Most of these are empty in a typical pattern set,
wasting a lot of memory. This change replaces the
internal representation with immutable sets, which
take up less memory.

This assumes that pattern sets are only mutated
infrequently. It also changes the behavior of the
`getIncludes` method and its peers. They now return
an immutable view. This is more in line with the rest
of the API, which suggests that methods like
`include(String...)` should be used instead of mutating
the sets returned from the getters.