Make task creation by Map cheaper
Task creation by map is used when the user types
the following in the DSL:

```
task foo(type:Bar) {}
```

It was always creating two intermediate maps,
which were only necessary for the case when
mandatory arguments were missing or when the
obscure `replace` flag was used. This change
avoids the creation of those maps when possible.