Prevent reallocation in ExtensibleDynamicObject
When the delegates of an extensible dynamic object
were changed, it would recompute its delegate array
and trim it to size. This trimming didn't save very
much. The reallocation on the other hand was leading
to a lot of avoidable garbage.

The delegate array is now fixed and can contain nulls
for delegates which are not yet set.